+++ browser_test/children_test.js
@@ -1,5 +1,5 @@
 /* eslint-env mocha */
-import { React, Remount, root, IS_DEBUG, raf, assert } from './setup'
+import { assert, raf } from './utils'
 
 const Dumper = props => {
   return <span className='dumper'>[{JSON.stringify(props)}]</span>
