+++ browser_test/test.js
@@ -1,4 +1,5 @@
 /* eslint-env mocha */
+import './setup'
 import './env_test'
 import './remount_test'
 import './inception_test'
